replaced taskControl to core and updated fields
