+++ apps/readmanga_parser/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 from apps.core.admin import AuthorLinkMixin, BaseAdmin, ImagePreviewMixin
-from apps.readmanga_parser.models import Author, Genre, Manga, TaskControl
+from apps.readmanga_parser.models import Author, Genre, Manga
 
 
 @admin.register(Manga)
@@ -31,9 +31,3 @@ class AuthorAdmin(BaseAdmin, admin.ModelAdmin):
 class GenreAdmin(BaseAdmin, admin.ModelAdmin):
     search_fields = ("name",)
     list_display = ("name",)
-
-
-@admin.register(TaskControl)
-class TaskControlAdmin(BaseAdmin, admin.ModelAdmin):
-    list_display = ("task_name", "task_status")
-    list_editable = ("task_status",)
